Add getMostRecentPosts and getMostCommentedPosts functionality to PostController

+++ ForumSystem/src/main/java/com/telerikacademy/web/forumsystem/models/Post.java
@@ -111,7 +111,7 @@ public class Post {
         return comments;
     }
 
-    public int getCommentsSize(){
+    public long getCommentsSize(){
         return comments.size();
     }
